Finalizado

+++ RestauranteFront/ClientApp/app/components/btn/btn-remove.component.ts
@@ -2,6 +2,7 @@
 import { ICellRendererAngularComp } from "ag-grid-angular";
 import { Router } from "@angular/router";
 import { AppService } from "../app.service";
+import swal from 'sweetalert2'
 
 @Component({
     selector: 'child-cell',
@@ -21,11 +22,49 @@ export class BtnRemoveComponent implements ICellRendererAngularComp {
         let url: string = this.router.url;
 
         if (url.indexOf("Restaurante") !== -1) {
-            this.service.RemoverRestaurante(this.params.node.data.RestauranteId);
-            this.refresh();
+            swal({
+                title: 'Tem certeza que deseja remover o restaurante?',
+                text: "A remoção é irreversível!",
+                type: 'warning',
+                showCancelButton: true,
+                confirmButtonColor: '#3085d6',
+                cancelButtonColor: '#d33',
+                cancelButtonText: "Não",
+                confirmButtonText: 'Sim, remova!'
+            }).then((result) => {
+                if (result.value) {
+
+                    this.service.RemoverRestaurante(this.params.node.data.RestauranteId);
+                    swal(
+                        'Removido!',
+                        'O restaurante foi removido com sucesso!',
+                        'success'
+                    )
+                }
+            });
         }
         else {
-            this.service.RemoverPrato(this.params.node.data.PratoId);
+            swal({
+                title: 'Tem certeza que deseja remover o prato?',
+                text: "A remoção é irreversível!",
+                type: 'warning',
+                showCancelButton: true,
+                confirmButtonColor: '#3085d6',
+                cancelButtonColor: '#d33',
+                cancelButtonText: "Não",
+                confirmButtonText: 'Sim, remova!'
+            }).then((result) => {
+                if (result.value) {
+
+                    this.service.RemoverPrato(this.params.node.data.PratoId);
+                    swal(
+                        'Removido!',
+                        'O prato foi removido com sucesso!',
+                        'success'
+                    )
+                }
+            });
+            
         }
     }
